Fix rank bounds and change original rank method
Wilcoxon W code looped one index too far.
Original ranking gave ties different rankings, and now ties share the
same (averaged) rank.

+++ World/NKWorld/NKWorld.h
@@ -21,7 +21,7 @@
 
 struct RankEpistasisData{
     size_t offset;
-    size_t rank;
+    double rank;
     double score_original;
     double score_mutant;
 };
